fix: memory reserve size underflow

+++ src/fastertransformer/core/TrackerAllocator.h
@@ -7,7 +7,7 @@ namespace fastertransformer {
 
 struct TrackerAllocatorParams {
     IAllocator* real_allocator = nullptr;
-    size_t target_track_bytes  = 0;
+    int64_t target_track_bytes = 0;
     size_t bytes_try_step      = 64UL * 1024 * 1024; // 64 MiB
     size_t align_size          = 1024;
 };
